test-requirements.txt
# https://github.com/OCA/account-financial-reporting/pull/1137
odoo-addon-account_tax_balance @ git+https://github.com/OCA/account-financial-reporting.git@refs/pull/1137/head#subdirectory=setup/account_tax_balance
